将JdbcDao改造成接口的模式，并启用动态代码的方式执行

+++ custom-db-action/src/main/java/com/custom/action/dbaction/JdbcActiveWrapper.java
@@ -43,7 +43,7 @@ public interface JdbcActiveWrapper<T, P> {
     int updateByCondition(T t, ConditionWrapper<T> wrapper);
 
     /*------------------------------------ comm ---------------------------------------*/
-    long save(T t);
+    int save(T t);
     P primaryKeyValue(T entity);
